requirements.txt
# API Framework
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.3.0

# Database
sqlalchemy>=2.0.20
alembic>=1.12.0
psycopg2-binary>=2.9.7  # For PostgreSQL
aiosqlite>=0.19.0  # For SQLite async support

# AI/ML
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.24.1
tenacity>=8.2.3  # For retries

# For development/testing
pytest>=7.4.0
pytest-asyncio>=0.21.1 